Moving most of Layer stuff to ImageLayer class
The current Layer class was designed to be an image layer itself. From
now on, it should be a base class for any types of layers to be created,
and as much abstract as possible.

Listing some other relevant change on Layer/ImageLayer class:

* "Layers" component has been removed. It was implemented to be a
  container of Layer objects, but it is no longer required. Layer
  objects must be direct childs of a Scene.
* Property "order" removed. If the declaration order needs to be
  overriden, the "z" property must be used instead.
* Property "factor" logic merged into "horizontalStep". This is used to
  set different horizontal animation (scroll) speed for a Layer.
  Negative values scroll to left and positive ones to right.
