+++ src/layer.cpp
@@ -30,402 +30,50 @@
 //! Class constructor
 Layer::Layer(QuasiDeclarativeItem *parent)
     : QuasiPaintedItem(parent)
-    , m_direction((Quasi::LayerDirection)-1) // Backward
-    , m_tileWidth(32)
-    , m_tileHeight(32)
-    , m_factor(1.0)
-    , m_drawType(Quasi::TiledDrawType)
+    , m_isAnimated(false)
+    , m_horizontalStep(1.0)
     , m_type(Quasi::InfiniteType)
-    , m_areaToDraw(2.0)
-    , m_columnOffset(0)
-    , m_latestPoint(0)
 {
 #if QT_VERSION >= 0x050000
-    setZ(Quasi::InteractionLayerOrdering_01);
     // this activates the item layered mode
     QQmlProperty(this, "layer.enabled").write(true);
-#else
-    setZValue(Quasi::InteractionLayerOrdering_01);
 #endif
 }
 
 //! Class destructor
 Layer::~Layer()
 {
-    m_pixmaps.clear();
 }
 
-//! Stores the source path for the image
-/*!
- * \param source the image path
- */
-void Layer::setSource(const QString &source)
+void Layer::setAnimated(bool animated)
 {
-    if (m_source != source)
-        m_source = source;
+    if (m_isAnimated == animated)
+        return;
+
+    m_isAnimated = animated;
+
+    emit animatedChanged();
 }
 
-//! Gets the image source path
-/*!
- * \return the source path for the image
- */
-QString Layer::source() const
+void Layer::setHorizontalStep(const qreal &step)
 {
-    return m_source;
-}
+    if (m_horizontalStep == step)
+        return;
 
-//! Stores the layer type
-/*!
- * \param drawType can be Tiled (default) or Plane
- */
-void Layer::setDrawType(Quasi::DrawType drawType)
-{
-    if (m_drawType != drawType)
-        m_drawType = drawType;
-}
+    if ((step > 0 && m_horizontalStep < 0) || (step < 0 && m_horizontalStep > 0))
+        emit horizontalDirectionChanged();
 
-//! Gets the layer type
-/*!
- * \return Tiled or Plane according the layer draw type
- */
-Quasi::DrawType Layer::drawType() const
-{
-    return m_drawType;
-}
+    m_horizontalStep = step;
 
-void Layer::setDirection(const Quasi::LayerDirection &direction)
-{
-    if (direction != m_direction) {
-        m_direction = direction;
-
-        if (m_offsets.count() != 0)
-            m_columnOffset = (m_columnOffset + 2) % m_offsets.size();
-
-        emit directionChanged();
-    }
-}
-
-//! Stores the layer update factor
-/*!
- * \param factor the factor value
- */
-void Layer::setFactor(qreal factor)
-{
-    if (m_factor != factor) {
-        m_factor = factor;
-
-        emit factorChanged();
-    }
-}
-
-//! Gets the layer update factor
-/*!
- * \return layer update factor
- */
-qreal Layer::factor() const
-{
-    return m_factor;
-}
-
-//! Stores the layer z order
-/*!
- * \param order the layer z order
- */
-void Layer::setOrder(Quasi::Ordering order)
-{
-#if QT_VERSION >= 0x050000
-    if (z() != order)
-        setZ(order);
-#else
-    if (zValue() != order)
-        setZValue(order);
-#endif
-}
-
-//! Gets the layer z order
-/*!
- * \return layer z order
- */
-Quasi::Ordering Layer::order() const
-{
-#if QT_VERSION >= 0x050000
-    return (Quasi::Ordering)z();
-#else
-    return (Quasi::Ordering)zValue();
-#endif
+    emit horizontalStepChanged();
 }
 
 void Layer::setLayerType(const Quasi::LayerType &type)
 {
-    if (type != m_type) {
-        m_type = type;
-
-        emit layerTypeChanged();
-    }
-}
-
-
-void Layer::setTileHeight(const int &value)
-{
-    if (m_drawType == Quasi::PlaneDrawType)
+    if (type == m_type)
         return;
 
-    if (value != m_tileHeight) {
-        m_tileHeight = value;
+    m_type = type;
 
-        if (m_tileWidth != 0 && m_tileHeight != 0)
-            emit tilesChanged();
-    }
-}
-
-void Layer::setTileWidth(const int &value)
-{
-    if (m_drawType == Quasi::PlaneDrawType)
-        return;
-
-    if (value != m_tileWidth) {
-        m_tileWidth = value;
-
-        if (m_tileWidth != 0 && m_tileHeight != 0)
-            emit tilesChanged();
-    }
-}
-
-//! Adds a tile on the list
-/*!
- * \param pix the pixmap to append on the list
- * \return the list actual size or -1 if the layer can not accept tiled pixmaps
- */
-int Layer::addTile(const QPixmap &pix)
-{
-    m_pixmaps.append(pix);
-
-    return m_pixmaps.size();
-}
-
-//! Gets a tile from the list
-/*!
- * \param pos the tile position on the list
- * \return the tile pixmap of position pos on the list or null, if none
- */
-QPixmap Layer::getTile(int pos) const
-{
-    return m_pixmaps.at(pos);
-}
-
-void Layer::setDrawGrid(bool draw)
-{
-    if (draw != m_drawGrid)
-        m_drawGrid = draw;
-}
-
-void Layer::setGridColor(const QColor &color)
-{
-    if (color != m_gridColor)
-        m_gridColor = color;
-}
-
-//! Gets the tiles pixmap list size
-/*!
- * \return the tiles pixmap list size
- */
-int Layer::count() const
-{
-    return m_pixmaps.size();
-}
-
-void Layer::generateOffsets()
-{
-    bool completed = false;
-    int start = 0;
-    int step = m_numColumns;
-    int max = m_totalColumns;
-    int count = 0;
-    int maxCount = step * (int)m_areaToDraw;
-    bool first = true;
-    Offsets::OffsetsList firstPoint;
-
-    while (!completed) {
-        Offsets::OffsetsList offsetsList;
-
-        int tamanho;
-        int fim = 0;
-        bool finish = false;
-
-        while (count < maxCount) {
-            fim = (start + step) % max;
-
-            if (fim - start > 0) {
-                tamanho = step;
-                count += tamanho;
-
-                // TODO check this comparison. Is it really needed?
-                if (finish || count != maxCount) {
-                    offsetsList.append(Offsets(start, tamanho));
-
-                    if (!finish)
-                        start = fim;
-                    finish = false;
-                } else {
-                    offsetsList.append(Offsets(start, tamanho));
-                }
-            } else {
-                int oldStart = start;
-                tamanho = max - start;
-                count += tamanho;
-
-                offsetsList.append(Offsets(start, tamanho));
-
-                tamanho = step - tamanho;
-                start = 0;
-                count += tamanho;
-
-                if (tamanho != 0) {
-                    offsetsList.append(Offsets(0, tamanho));
-                }
-
-                if (count <= maxCount / 2) {
-                    start = tamanho;
-                    finish = true;
-                } else
-                    start = oldStart;
-            }
-        }
-
-        count = 0;
-
-        if (offsetsList == firstPoint)
-            completed = true;
-        else
-            m_offsets.append(offsetsList);
-
-        if (first) {
-            firstPoint = offsetsList;
-            first = false;
-        }
-    }
-}
-
-void Layer::updateTiles()
-{
-    if ((boundingRect().width() == 0) || (boundingRect().height() == 0))
-        return;
-
-    // TODO create enums to define image aspect, auto tile, etc...
-    QPixmap pix(source()); // TODO
-
-    if (m_drawType == Quasi::PlaneDrawType) {
-        m_tileWidth = width();
-        m_tileHeight = height();
-
-        if (pix.width() % (int)width() != 0) {
-            // XXX create some log system?
-            qCritical() << QString("Quasi>>Image \'%1\' doesn't contains a proper size... CROPPING!").arg(source());
-
-            int newWidth = pix.width() - (pix.width() % (int)width());
-            pix = pix.copy(0, 0, newWidth, height());
-        }
-    }
-
-    if (pix.width() < boundingRect().width()) {
-        QPixmap temp(boundingRect().width(), boundingRect().height());
-        QPainter p(&temp);
-            p.drawTiledPixmap(boundingRect(), pix, QPoint(0,0));
-        p.end();
-
-        pix = temp;
-    }
-
-    if (m_type == Quasi::MirroredType) {
-        QPixmap temp(pix.width() * 2, pix.height());
-
-        QPainter p(&temp);
-            p.drawPixmap(0, 0, pix.width(), pix.height(), pix);
-            p.drawPixmap(pix.width(), 0, pix.width(), pix.height(),
-                        pix.transformed(QTransform().scale(-1, 1), Qt::FastTransformation));
-        p.end();
-
-        pix = temp;
-    }
-
-    // visible tiles
-    m_numColumns = boundingRect().width() / m_tileWidth;
-    m_numRows = boundingRect().height() / m_tileHeight;
-
-    // total of columns and rows
-    m_totalColumns = pix.width() / m_tileWidth;
-    m_totalRows = pix.height() / m_tileHeight;
-
-    int i, j;
-    for (i = 0; i < m_totalRows; i++) {
-        for (j = 0; j < m_totalColumns; j++) {
-            QPixmap temp(m_tileWidth, m_tileHeight);
-
-            QPainter p(&temp);
-                p.setCompositionMode(QPainter::CompositionMode_Source);
-                p.drawPixmap(0, 0, m_tileWidth, m_tileHeight,
-                        pix, j * m_tileWidth, i * m_tileHeight, m_tileWidth, m_tileHeight);
-            p.end();
-
-            addTile(temp);
-        }
-    }
-
-    generateOffsets();
-    drawPixmap();
-}
-
-QPixmap Layer::generatePartialPixmap(int startPoint, int size)
-{
-    QPixmap temp(m_tileWidth * size, boundingRect().height());
-
-    QPainter p(&temp);
-        int i, j;
-        int index = 0;
-        for (i = 0; i < m_numRows; i++) {
-            for (j = 0; j < size; j++) {
-                index = ((i * m_totalColumns) + (j + startPoint));
-
-                p.drawPixmap(j * m_tileWidth, i * m_tileHeight, getTile(index));
-
-                // just draw a grid
-                // XXX chech the possibility of drawn it only on a debug mode
-                if (m_drawGrid) {
-                    p.setPen(m_gridColor);
-                    p.drawRect(j * m_tileWidth, i * m_tileHeight, m_tileWidth, m_tileHeight);
-                }
-            }
-        }
-    p.end();
-
-    return temp;
-}
-
-void Layer::drawPixmap()
-{
-    if ((boundingRect().width() == 0) || (boundingRect().height() == 0))
-        return;
-
-    if (m_currentImage)
-        delete m_currentImage;
-
-    m_currentImage = new QImage(boundingRect().width() * m_areaToDraw, boundingRect().height(), QImage::Format_ARGB32_Premultiplied);
-
-    QPainter p(m_currentImage);
-        int xPoint = 0;
-        for (int i = 0; i < m_offsets[m_columnOffset].size(); i++) {
-            Offsets offset = m_offsets[m_columnOffset].at(i);
-
-            QPixmap pix = generatePartialPixmap(offset.point(), offset.size());
-            p.drawPixmap(xPoint, 0, pix);
-
-            xPoint += pix.width();
-            m_latestPoint = offset.point();
-        }
-
-        if (m_direction == Quasi::ForwardDirection)
-            m_columnOffset = (m_columnOffset - 1 < 0) ? m_offsets.size() - 1 : m_columnOffset - 1;
-        else
-            m_columnOffset = (m_columnOffset + 1) % m_offsets.size();
-    p.end();
+    emit layerTypeChanged();
 }
